+++ lucifer/tests/scripts/check_coherent.py
@@ -83,7 +83,7 @@ NML = """&fel_params
   wavefront_init%lambda0 = 1e-10
   wavefront_init%seed_power = 5e3
   wavefront_init%seed_waist_size = 30e-6
-  wavefront_init%grid_n_pts = 151
+  wavefront_init%grid_n_pts = 128
   wavefront_init%grid_half_width = 2e-4
 {wextra}/
 """